refactor(core-api): rename the file of `buildFilterQuery` utility
